Split up the super-large Variables.cc file
