+++ opencog/atoms/core/Replacement.h
@@ -1,0 +1,64 @@
+/*
+ * opencog/atoms/core/Replacement.h
+ *
+ * Copyright (C) 2015 Linas Vepstas
+ * All Rights Reserved
+ *
+ * This program is free software; you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License v3 as
+ * published by the Free Software Foundation and including the exceptions
+ * at http://opencog.org/wiki/Licenses
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program; if not, write to:
+ * Free Software Foundation, Inc.,
+ * 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
+ */
+
+#ifndef _OPENCOG_REPLACEMENT_H
+#define _OPENCOG_REPLACEMENT_H
+
+#include <map>
+#include <set>
+
+#include <boost/operators.hpp>
+
+#include <opencog/util/empty_string.h>
+#include <opencog/atoms/base/Handle.h>
+#include <opencog/atoms/base/Atom.h>
+#include <opencog/atoms/core/Quotation.h>
+
+namespace opencog
+{
+/** \addtogroup grp_atomspace
+ *  @{
+ */
+
+/// Given a tree, replace one Atom by another in that tree, respecting
+/// quotation (QuoteLink) and scoping (ScopeLink).
+struct Replacement
+{
+	typedef std::map<Handle, unsigned int> IndexMap;
+
+	/// Walk the tree given in the first argument, and replace
+	/// any atoms that occur in the map by thier mapped value.
+	static Handle replace_nocheck(const Handle&, const HandleMap&);
+
+protected:
+	static Handle substitute_scoped(Handle, const HandleSeq&,
+	                                const IndexMap&,
+	                                Quotation quotation=Quotation());
+	static bool must_alpha_convert(const Handle& scope, const HandleSeq& args);
+	static bool must_alpha_hide(const Handle& scope, const IndexMap& index_map);
+	static IndexMap alpha_hide(const Handle& scope, const IndexMap& index_map);
+};
+
+/** @}*/
+}
+
+#endif // _OPENCOG_REPLACEMENT_H
